chore: add backgrounds to docs
